+++ items.lua
@@ -11,4 +11,10 @@ PlaceObj('ModItemCode', {
 	'name', "AT_3ModConfig",
 	'FileName', "Code/AT_3ModConfig.lua",
 }),
+PlaceObj('ModItemDecalEntity', {
+	'name', "Mars Touring Company",
+	'comment', "Mars Touring - Better than Uranus",
+	'entity_name', "AutomatedTourismLogo",
+	'filename', "UI/Logos/TourCompanyLogo.png",
+}),
 }